pyquizgen/pyproject.toml
[tool.poetry]
name = "pyquizgen"
version = "0.1.0"
description = ""
authors = ["marcacohen <marcacohen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.240"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
requests-mock = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
